Cache collisions by positions instead of delta
The collision cache is used to avoid notifying of a collision which is
invisible to the player, i.e. when Pickle is pushed up onto the floor.

This was originally done by caching the velocity delta between two
objects, which was susceptible to a bug where the collision would not
notify identical deltas if there was movement in between (i.e. Pickle
jumped and then landed with the same velocity delta). Using a
position-based approach more accurately determines if a collision is
visually distinct or if the object was resolved to its previous position